+++ packages/canonical-bridge-widget/src/modules/aggregator/hooks/useSelection.ts
@@ -312,6 +312,7 @@ function useSortedTokens() {
           chainId: fromChainId,
           chainType,
           tokenAddress: item?.address,
+          tokenSymbol: item?.symbol,
         });
 
         let value: number | undefined;